Replace annotation-sources.txt with git submodules
This is just a suggestion. Instead of pulling the annotation sources from
python they could also be added to the repository as git submodules. You need
less of your own code (so less that can go wrong there), and allows working
within the git ecosystem in the with git convential ways (such as adding more
submodules in your own fork etc.)